+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/adapter/FriendAdapter.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation.main.home
+package org.sopt.dosopttemplate.presentation.main.home.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
