+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/adapter/FriendMusicViewHolder.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation.main.home
+package org.sopt.dosopttemplate.presentation.main.home.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.dosopttemplate.databinding.ItemFriendMusicBinding
